Moved the whitelist functions out of the AuthHandler to their own class
